Replace more int32 with int64 in folder 'integrationTests'

+++ integrationTests/mock/scProcessorMock.go
@@ -9,8 +9,8 @@ import (
 
 type SCProcessorMock struct {
 	ComputeTransactionTypeCalled          func(tx *transaction.Transaction) (process.TransactionType, error)
-	ExecuteSmartContractTransactionCalled func(tx *transaction.Transaction, acntSrc, acntDst state.AccountHandler, round uint32) ([]*smartContractResult.SmartContractResult, error)
-	DeploySmartContractCalled             func(tx *transaction.Transaction, acntSrc state.AccountHandler, round uint32) ([]*smartContractResult.SmartContractResult, error)
+	ExecuteSmartContractTransactionCalled func(tx *transaction.Transaction, acntSrc, acntDst state.AccountHandler, round uint64) ([]*smartContractResult.SmartContractResult, error)
+	DeploySmartContractCalled             func(tx *transaction.Transaction, acntSrc state.AccountHandler, round uint64) ([]*smartContractResult.SmartContractResult, error)
 	ProcessSmartContractResultCalled      func(scr *smartContractResult.SmartContractResult) error
 }
 
@@ -27,7 +27,7 @@ func (sc *SCProcessorMock) ComputeTransactionType(
 func (sc *SCProcessorMock) ExecuteSmartContractTransaction(
 	tx *transaction.Transaction,
 	acntSrc, acntDst state.AccountHandler,
-	round uint32,
+	round uint64,
 ) ([]*smartContractResult.SmartContractResult, error) {
 	if sc.ExecuteSmartContractTransactionCalled == nil {
 		return nil, nil
@@ -39,7 +39,7 @@ func (sc *SCProcessorMock) ExecuteSmartContractTransaction(
 func (sc *SCProcessorMock) DeploySmartContract(
 	tx *transaction.Transaction,
 	acntSrc state.AccountHandler,
-	round uint32,
+	round uint64,
 ) ([]*smartContractResult.SmartContractResult, error) {
 	if sc.DeploySmartContractCalled == nil {
 		return nil, nil
